--- main
@@ -1,2 +1,0 @@
-collection = Collection.new
-collection.run
